+++ pkg/sql/colflow/colrpc/colrpc_test.go
@@ -286,14 +286,14 @@ func TestOutboxInbox(t *testing.T) {
 				if outputBatch == coldata.ZeroBatch {
 					outputBatches.Add(coldata.ZeroBatch)
 				} else {
-					batchCopy := testAllocator.NewMemBatchWithSize(typs, int(outputBatch.Length()))
+					batchCopy := testAllocator.NewMemBatchWithSize(typs, outputBatch.Length())
 					testAllocator.PerformOperation(batchCopy.ColVecs(), func() {
 						for i := range typs {
 							batchCopy.ColVec(i).Append(
 								coldata.SliceArgs{
 									ColType:   typs[i],
 									Src:       outputBatch.ColVec(i),
-									SrcEndIdx: uint64(outputBatch.Length()),
+									SrcEndIdx: outputBatch.Length(),
 								},
 							)
 						}
@@ -334,8 +334,8 @@ func TestOutboxInbox(t *testing.T) {
 				for i := range typs {
 					require.Equal(
 						t,
-						inputBatch.ColVec(i).Window(typs[i], 0, uint64(inputBatch.Length())),
-						outputBatch.ColVec(i).Window(typs[i], 0, uint64(outputBatch.Length())),
+						inputBatch.ColVec(i).Window(typs[i], 0, inputBatch.Length()),
+						outputBatch.ColVec(i).Window(typs[i], 0, outputBatch.Length()),
 						"batchNum: %d", batchNum,
 					)
 				}
